% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignments}
\alias{assignments}
\title{Create an assignment document with NMU styling}
\usage{
assignments(..., keep_tex = TRUE, highlight = "monochrome")
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}

\item{keep_tex}{A boolean toggle to select whether intermediate
LaTeX files are to be kept, defaults to \code{TRUE}}

\item{highlight}{Syntax highlighting style, defaults to "monochrome"}
}
\value{
R Markdown output format to pass to render
}
\description{
Create an assignment document with NMU styling
}
