% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_base_pdf_format}
\alias{create_base_pdf_format}
\title{Base PDF format creator}
\usage{
create_base_pdf_format(
  template_path,
  keep_tex = TRUE,
  includes = NULL,
  pandoc_args = NULL,
  ...
)
}
\description{
Base PDF format creator
}
\keyword{internal}
