% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer.R
\name{nmu_beamer}
\alias{nmu_beamer}
\title{Create NMU Beamer Presentation}
\usage{
nmu_beamer(..., keep_tex = FALSE, theme = "sciences")
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:beamer_presentation]{rmarkdown::beamer_presentation()}}}

\item{keep_tex}{Keep generated LaTeX output}

\item{theme}{The NMU faculty theme to use (one of "humanities", "education",
"business", "sciences", "health", "engineering", "law", "ocean")}
}
\value{
R Markdown output format to pass to render
}
\description{
Create NMU Beamer Presentation
}
