% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prac.R
\name{pdf_prac}
\alias{pdf_prac}
\title{Render PDF Document With NMU Format}
\usage{
pdf_prac(..., keep_tex = FALSE)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}

\item{keep_tex}{A boolean toggle to select whether intermediate
LaTeX files are to be kept, defaults to \code{FALSE}}
}
\value{
R Markdown output format to pass to render.
}
\description{
Render PDF Document With NMU Format
}
