% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_notes.R
\name{short_notes}
\alias{short_notes}
\title{Render PDF Document With NMU Short Notes Format}
\usage{
short_notes(..., keep_tex = FALSE)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}

\item{keep_tex}{A boolean toggle to select whether intermediate
LaTeX files are to be kept, defaults to \code{FALSE}}
}
\value{
R Markdown output format to pass to render.
}
\description{
Render PDF Document With NMU Short Notes Format
}
