% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prac.R
\name{html_prac}
\alias{html_prac}
\title{Render HTML Document With NMU Format}
\usage{
html_prac(toc = FALSE)
}
\arguments{
\item{toc}{Logical, include table of contents. Default is FALSE.}
}
\value{
R Markdown output format to pass to render.
}
\description{
Render HTML Document With NMU Format
}
